+++ DatadogRUM/Tests/Mocks/RUMDataModelMocks.swift
@@ -143,7 +143,7 @@ extension RUMViewEvent: RandomMockable {
                 pageStates: nil,
                 replayStats: nil,
                 session: .init(
-                    plan: [.plan1, .plan2].randomElement()!,
+                    plan: .plan1,
                     sessionPrecondition: .mockRandom()
                 )
             ),
@@ -442,19 +442,6 @@ extension RUMErrorEvent: RandomMockable {
                 url: .mockRandom()
             )
         )
-    }
-}
-
-extension RUMCrashEvent: RandomMockable {
-    static func mockRandom(error: RUMErrorEvent) -> RUMCrashEvent {
-        return .init(
-            error: error,
-            additionalAttributes: mockRandomAttributes()
-        )
-    }
-
-    public static func mockRandom() -> RUMCrashEvent {
-        return mockRandom(error: .mockRandom())
     }
 }