+++ DatadogTrace/Sources/Feature/TraceFeature.swift
@@ -30,7 +30,8 @@ internal final class TraceFeature: DatadogRemoteFeature {
             core: core,
             sampler: Sampler(samplingRate: configuration.debugSDK ? 100 : configuration.sampleRate),
             tags: configuration.tags ?? [:],
-            tracingUUIDGenerator: configuration.traceIDGenerator,
+            traceIDGenerator: configuration.traceIDGenerator,
+            spanIDGenerator: configuration.spanIDGenerator,
             dateProvider: configuration.dateProvider,
             loggingIntegration: TracingWithLoggingIntegration(
                 core: core,